doc/misc: add LoadBalancer Guide to docs.ceph.com
This commit adds the LoadBalancer Guide to docs.ceph.com.
In an effort to encourage more community contributions to
Ceph documentation, a third-party-utilities page will
eventually be created so that Ceph-adjacent technologies
have a place to be discovered on the docs.ceph.com website.
This is the first such third-party utility to be added.
Since it is only the first, it is linked from the main
installation page, but as more third-party Ceph-adjacent
utilities are included in the website, this link will move
to a dedicated third-party-utilities page.

Fixes: https://tracker.ceph.com/issues/46421
Signed-off-by: Zac Dover <zac.dover@gmail.com>