mgr/dashboard: Fix CephFS's Directories not showing
This is a regretion introduced in 5a9d6d0bad7eef0e6e7f0b76c4bac85b3cf35041.

We no longer need to check if a tab is selected to load the content.

In that commit we removed the logic that updated the conditional variable,
but  missed removing its use.

Fixes: https://tracker.ceph.com/issues/45877

Signed-off-by: Tiago Melo <tmelo@suse.com>