Fix incorrect tracking of caps per session by the client
When a cap is removed via remove_cap(), the cap is set to nullptr but that does not mean that it is removed from session->caps. As of now the cap is removed from session-caps only when the destructor of Cap is called and this happens when the inode is deleted (i.e during trimming of inode).

Ideally we should remove the cap structure from the session->caps list whenever remove_cap() is called.

Fixes: https://tracker.ceph.com/issues/45749
Signed-off-by: Sidharth Anupkrishnan <sanupkri@redhat.com>