cephadm: make 'ls' faster
Instead of getting the info for the 'ls' command by inspecting
each container it can mostly be pulled from docker/podman 'ps'
and 'image list/ls' commands with the exception of the version.
This should make it faster when run on hosts with many containers
built using the same image.

Fixes: https://tracker.ceph.com/issues/45055
Signed-off-by: Adam King <adking@redhat.com>