doc/rados: Update erasure code profile guidance
Add a plain language and a mathematical explanation of K and M and their relation to overhead.
Add an explanation that the minimum pool size is K+M of the failure domain. Add guidance to
include one more failure domain than the minimum to remain healthy under partial failure (K+M+1).

Signed-off-by: Kane York <kanepyork@gmail.com>