mds: reset sequence number when new rank0 mds is available
`MetricsHandler::last_updated_seq` is used to figure out if
client disconnection to an MDS needs to be sent to rank0 MDS
an as (remove) update. If a client establishes a session and
quickly tears it down before active MDSs can send a metric
update message to rank0 MDS, the remove update is never sent
to rank0.

There is also a case where a new rank0 recieves a metric remove
update for a session it never saw. This can happen when a metric
update was sent to an earlier rank0, followed by a failover to
a new rank0 and a client session disconnection.

Fixes: http://tracker.ceph.com/issues/46023
Signed-off-by: Venky Shankar <vshankar@redhat.com>