vstart: just use ganesha.nfsd that's in $PATH
In an earlier patch, I added a --ganesha-path option, but that left out
ganesha-rados-grace, which also could be in an alternate path. Instead,
just change vstart to run whatever ganesha.nfsd it finds in $PATH. We
already do that for ganesha-rados-grace so it should be fine to do that
for ganesha.nfsd as well.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit eab8c8136dc5379efde28e62f195437997c9db7f)

Conflicts:
	src/vstart.sh

- Removed bluestore-zoned option as pr(#35111) is not backported yet.