os/bluestore: Track per-zone metadata
We maintain a string per zone describing operations happening to the zone.  When
an object is written to the zone, the string is appended an "add object"
operation flag together with the object key length and the object key.  When a
copy of an object already in the zone is written elsewhere, the string is
appended an "invalidate object" operation flag together with the object key
length and the object key.  The cleaner will process this string to know the
object keys of the objects currently alive in the zone.

Signed-off-by: Abutalib Aghayev <agayev@cs.cmu.edu>