worker: fix crash on try_delete_jobs
In some circumstances workers try to delete a job which
is not created yet. It quits immediately with unhandled
exception error.

We do not want to allow crashes whatever error happens,
because worker can continue working and process other
jobs correctly.

Also some schedulers might not add corresponding
first in suite and last in suite jobs to the reporter,
since it is not required.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>