mon: elector: include the election strategy in MMonElection and check it
We saw a crash when we incorrectly assumed an MMonElection included
a scoring_bl, apparently because there was a disagreement on strategy.
This probably means the proposer was on an old MonMap (which we actually
DO NOT filter on when receiving proposals, oddly enough...), so now
we drop the message if there's a conflict. This will sort itself out
quickly enough!

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit b89b5438e8638fa27eeb17b7e543267873e1880b)

Conflicts:
	src/messages/MMonElection.h
Resolves: rhbz#1800382