mds: do not decode EMetaBlob::fullbit struct_v < 8
the struct_v of EMetaBlob::fullbit::encode was bumped from 7 to 8 in
fe286649822edfed8f7886a3ff0b1d8146396deb, the oldest release containing
this change was hammer rc1.

`EMetaBlob::fullbit` is used to persist dentry + inode as part of
journal, after the journal is flushed to the metadata pool, and applied to
dirty metadata, the flushed segments are trimmed. also, it's not allowed
to upgrade a Ceph cluster across 2 releases. so, there is no need
to decode an on-disk struct which will be trimmmed as part of journal
recycling, and that struct was encoded in hammer.

Signed-off-by: Kefu Chai <kchai@redhat.com>