mgr/volumes: Fix incorrect snapshot size
The size of the snapshot persisted in "ceph.dir.rbytes"
gets updated when I/O happens on parent subvolume. Hence
becomes stale. This patch persists the snapshot size in
new xattr "trusted.snapdir.rbytes" and provides correct size
through out the life time of snapshot.

Note that the size might be inconsistent if the I/O
happens while the snapshot is being taken. The size is also
inconsistent if "ceph.dir.rbytes" is not updated after the
I/O.

Fixes: https://tracker.ceph.com/issues/46158
Signed-off-by: Kotresh HR <khiremat@redhat.com>