mon: osd: add pg_pool_t::peering_crush_bucket_target, to ease peering/recovery
This will let us set a size of 4 and a target bucket count of 2, without
the OSDs trying to shove 4 replicas into a single bucket when in degraded
stretch mode. This also means we can stop changing the pool size
at all -- a "bucket_max" concept in calc_replicated_acting() will
prevent it from overfilling any given bucket in most cases.

That matters because CRUSH will frequently change the mapping
from PG->OSD when you change the size, depending on the rule construction.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>