librbd: fix parent cache races and error handling
If the plugin fails to connect to the daemon at start-up it will
crash the process due to a resource deadlock exception being
thrown as the client is destroyed. Additionally, librbd will support
concurrent IO thread processing in the future so the client needs
to be protected by a lock.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>