osd: conditionally encode stretch state in pg_pool_t, do not require new compat
This struct gets sent out to clients as well as OSDs, and we need them
to be able to decode it without crashing/failing. So we can't require it,
and happily the OSDs which might act on it will be gated by the OSDMap.

Additionally, we don't want older servers to fail OSDMap crc checks, so don't
encode the stretch data members if they aren't in use.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>