doc/cephadm: add octopus install guide
This commit adds the crisp Octopus Ceph Installation Guide
to the documentation. I have also added some links on the
landing page: one link points to this new guide and the
other points to the Developers' Guide. (A third link, to the
Ceph Bug Tracker, will follow in a future commit, which will
likely be associated with a different PR.)

Signed-off-by: Zac Dover <zac.dover@gmail.com>