doc/rados: Fix osd_op_queue default value
Since 14adc9d33f, `osd_op_queue` defaults to `wpq`, but the documentation was
still stating that its default value is `prio`.

This isn't being cherry-picked from master because it was fixed in be4582c26a
as part of a changed that modified the default value of a different option
(#30441).

Fixes: https://tracker.ceph.com/issues/46760
Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>