mgr/dashboard: right-align dropdown menu of column filters
When the value of a filter is too long, the text of value will run out
of the viewport. Right-aligning the menu makes the text visible.

NOTE: This change can't be backported from master because in master we had
switched the bootstrap library to ng-bootstrap and the menu is already
right-aligned in the master.

Fixes: https://tracker.ceph.com/issues/44458
Signed-off-by: Kiefer Chang <kiefer.chang@suse.com>