messages/MPGStats: print seq instead of tid
osd_stat.seq is a monotonic increasing sequence number which marks
the version of the of the report. since we don't use the tid to track
a MPGStats message and its MPGStatsAck message, it's more interesting to
us to print the sequence id. we use it to verify if a pg_stats is
flushed or not from a certain OSD.

Signed-off-by: Kefu Chai <kchai@redhat.com>