doc/index.rst: add quick-links
This is the first of a series of commits
meant to make the website more serendipitiously
browseable.

This commit adds three buttons at the top of
the landing page: (1) a link to the Octopus
Quick Start Guide, (2) a link to the
Developer Guide, and (3) a link to the Bug
Tracker. I have also written a small amount
of explanatory, signposting text. Another
PR will follow this one, which, inshallah,
will render the Developers Guide more
easily navigated.

Signed-off-by: Zac Dover <zac.dover@gmail.com>