msgr: interceptor test step enum to ease readability
I've been reading messenger code and tests to help familarise myself
with it. The test Interceptor is a great idea and gives breakpoint and
action control for simulating different connection situations. But found
the breakpoint step numbers confusing to read. These breakpoint steps do
not completely correspond to steps, as they're placed in useful places
for testing.

This patch wraps the numbers up in an enum inside the Interceptor struct
then this is used to replace the numbers to make the code more readable.

I could probably extend this to using the same Interceptor::STEP enums
in ProtocolV2.cc but as an initial patch I just wanted to touch the test
code.
Though maybe I'll do that anyway, because there's no point only making
it 1/2 readable ;)

The Interceptor::STEP enum names I just came up with based on where they
turn up in ProtocolV2, please feel free to suggest better ones.

Signed-off-by: Matthew Oliver <moliver@suse.com>