mon: store mon updates in ceph context for future MonMap instantiation
MonMap builds initial mon list using provided sources, like
mon-host or monmap.

For future instantiations of MonClient, if mon addresses are
updated, stale information from the provided sources are used.

This commit retains mon updates that are processed by the
MonClient in CephContext, for use in MonMap instantiations
and hence uses updated information as required.

This is helpful in cases where librados or libcephfs
instantiate MonClient in the ceph-mgr deamon as required.

Fixes: https://tracker.ceph.com/issues/46645
Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>