mon: mark pgtemp messages as no_reply more consistently in preprocess_pgtemp
If a message is forwarded, it's conceivable the leader's and peon's evaluation
will disagree about whether the message is useful or not, which could result
in the leader ignoring it and the peon having a dangling forwarded message.
Fix this by marking the op as no_reply whenever ignoring it.

Fixes:  https://tracker.ceph.com/issues/46914

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 73a014fc2ca928eb72def31c9e4177063cda421a)